Rename units.py to unit.py to avoid name clashes
